// backend/src/modules/links/links.schema.ts
const tag = z.object({
    name: z.string().min(1).max(30),
    color: z.string().regex(/^#[0-9a-fA-F]{6}$/).default('#6366f1'),
});

export const createLink = z.object({
    body: z.object({
        longUrl: z.string().url('Invalid URL').max(2048),
        customAlias: z
            .string()
            .min(3)
            .max(64)
            .regex(/^[a-zA-Z0-9_-]+$/, 'Alias can only contain letters, numbers, hyphens and underscores')
            .optional(),
        isOneTime: z.boolean().default(false).optional(),
        title: z.string().max(255).optional(),
        description: z.string().max(1000).optional(),
        passwordHash: z.string().min(4).max(100).optional(),
        expiresAt: z.string().datetime().optional(),
        clickLimit: z.number().int().min(1).max(1000000).optional(),
        redirectType: z.enum(['TEMPORARY', 'PERMANENT']).default('TEMPORARY'),
        tags: z.array(tag).max(10).optional(),
    })

});

export const updateLink = z.object({
    params: z.object({
        id: z.string().uuid()
    }),
    body: z.object({
        longUrl: z.string().url().max(2048).optional(),
        title: z.string().max(255).nullable().optional(),
        description: z.string().max(1000).nullable().optional(),
        isOneTime: z.boolean().optional(),
        customAlias: z.string()
            .min(3)
            .max(64)
            .regex(/^[a-zA-Z0-9_-]+$/, 'Alias can only contain letters, numbers, hyphens and underscores')
            .nullable()
            .optional(),
        password: z.string().min(4).max(100).nullable().optional(),
        expiresAt: z.string().datetime().nullable().optional(),
        clickLimit: z.number().int().min(1).optional(),
        redirectType: z.enum(['TEMPORARY', 'PERMANENT']).optional(),
        isActive: z.boolean().default(true).optional(),
        tags: z.array(tag).max(10).optional(),
    })
});
// ...
export const redirectRule = z.object({
    params: z.object({
        id: z.string().uuid()
    }),
    body: z.object({
        rules: z.array(
            z.object({
                conditionType: z.enum(['DEVICE', 'GEO', 'LANGUAGE', 'TIME_OF_DAY', 'DAY_OF_WEEK', 'DATE_RANGE']),
                conditionValue: z.string(), // JSON
                targetUrl: z.string().url(),
                label: z.string().max(100).optional(),
                priority: z.number().int().min(0).max(100).default(0),
                isActive: z.boolean().default(true),
            }),
        ).max(20),
    })
});
// ...
export type CreateLinkInput = z.infer<typeof createLink>['body'];
export type UpdateLinkInput = z.infer<typeof updateLink>['body'];
export type GetLinksQuery = z.infer<typeof getLinks>['query'];
export type BulkCreateInput = z.infer<typeof bulkCreate>['body'];

export type RedirectLinkIdInput = z.infer<typeof redirectRule>['params'];
export type RedirectRuleInput = z.infer<typeof redirectRule>['body'];